Revert "Start breaking out js stuff (#20)"
This reverts commit ec2c94968ae551aaabfa49bd78519d434bd84a82.
