+++ octoprint_usb_keyboard/static/js/knockout-file-bind.js
