--- octoprint_usb_keyboard/static/js/viewmodels/commands/command/octoprint_viewmodel.js
@@ -1,20 +1,0 @@
-function CommandsCommandOctoprintViewModel(params) {
-  var self = this
-  Lockable.call(self, "action", params.locked)
-  SelfManaged.call(self, params.parentArray, params.commandActionObject)
-  ShowsInfo.call(self)
-
-
-  // console.log("CommandsCommandOctoprintViewModel raw", params)
-  // console.log("CommandsCommandOctoprintViewModel self", self)
-
-  self.profile = params.profile;
-  self.type = params.commandActionObject.type;
-  self.command = params.commandActionObject.command;
-  self.presses_required = params.commandActionObject.presses_required.extend({ numeric: 0 });
-  self.supportedCommands = ["cancel_print", "confirm_last_command", "pause_print", "resume_print", "start_print", "toggle_pause_print"]
-}
-ko.components.register('sfr-commands-command-octoprint', {
-  viewModel: CommandsCommandOctoprintViewModel,
-  template: { element: 'template-sfr-commands-command-octoprint' }
-});
